+++ NEWS.md
@@ -1,3 +1,5 @@
+# probably (development version)
+
 # probably 1.0.0
 
 * Copyright holder changed to Posit Software PBC.