+++ NEWS.md
@@ -1,8 +1,10 @@
 # probably (development version)
 
-* `predict.int_conformal_infer_cv()` now returns a `.pred` column that is the average prediction from the resampled models. The prediction intervals are centered on these.
+* The conformal functions `int_conformal_infer_*()` were renamed to `int_conformal_*()`.
 
-* Split conformal inference is available using `int_conformal_infer_split()`. 
+* `predict.int_conformal_cv()` now returns a `.pred` column that is the average prediction from the resampled models. The prediction intervals are centered on these.
+
+* Split conformal inference is available using `int_conformal_split()`. 
 
 # probably 1.0.0